% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{has_meogis}
\alias{has_meogis}
\title{Test if the root path can be found.}
\usage{
has_meogis(x, root = options("MEOGIS_PATH"))
}
\arguments{
\item{x}{subdirectory (ala "lowtide2013" etc)}

\item{root}{the path to the various MEOGIS datasets}
}
\value{
named logical, TRUE if the path exists
}
\description{
The function attempts to retrieve this from option("MEOGIS_PATH") which can be
set at the begining of the R session or placed in the users \code{~\.Rprofile}
}
