% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meogis-package.R
\docType{package}
\name{meogis-package}
\alias{meogis-package}
\alias{meogis}
\title{meogis package}
\description{
Access, query and download data from Maine Office of GIS
}
\details{
Access Maine Office of GIS data files (local or online).  Datasets are stored in
 geojson format and loaded easily into simple features (sf).  These points
 and polygons can be manipulated/filtered using dplyr and sf packages.
}
\seealso{
\url{https://www.maine.gov/megis/catalog/}
}
