% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{meogis_path}
\alias{meogis_path}
\title{Retrieve the MEOGIS path}
\usage{
meogis_path(...)
}
\arguments{
\item{...}{further arguments for \code{has_nsspme}}
}
\value{
the known MEOGIS path
}
\description{
Retrieve the MEOGIS path
}
