% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_twp24}
\alias{read_twp24}
\title{Read the Maine Township 24k Polygons}
\usage{
read_twp24(filename = "METWP24.geojson", path = meogis_path(), ...)
}
\arguments{
\item{filename}{the name of the file, defaults to \code{METWP24.geojson}}

\item{path}{the path to the data file, defaults to MEOGIS_PATH}

\item{...}{further arguments for \code{sf::read_sf()}}
}
\value{
sf-tibble or NULL
}
\description{
Read the Maine Township 24k Polygons
}
